chore/feat: cleaned up views, added quiz_results More views have been cleaned up and moved into respective properties and helper functions instead of inline, also added results page
